+++ arangod/Aql/Optimizer/Rule/BatchMaterializeDocuments.cpp
