+++ arangod/Aql/Optimizer/Rule/CMakeLists.txt
@@ -1,20 +1,20 @@
 add_subdirectory(EnumeratePathsFilter)
 
 target_sources(arango_aql PRIVATE
-  OptimizerRuleBatchMaterializeDocuments.cpp
-  OptimizerRuleCollectWithIndex.cpp
-  OptimizerRulePropagateConstantAttributes.cpp
-  OptimizerRulePushDownLateMaterialization.cpp
-  OptimizerRulePushLimitIntoIndex.cpp
-  OptimizerRuleReplaceEntriesWithObjectIteration.cpp
-  OptimizerRuleReplaceEqualAttributeAccess.cpp
-  OptimizerRuleVectorIndex.cpp
-  OptimizerRulePushFilterVectorIndex.cpp
-  OptimizerRulesCluster.cpp
-  OptimizerRulesGraph.cpp
-  OptimizerRulesIResearchView.cpp
-  OptimizerRulesIndexNode.cpp
-  OptimizerRulesJoin.cpp
-  OptimizerRulesLateMaterializedCommon.cpp
-  OptimizerRulesReplaceFunctions.cpp
+  BatchMaterializeDocuments.cpp
+  CollectWithIndex.cpp
+  PropagateConstantAttributes.cpp
+  PushDownLateMaterialization.cpp
+  PushLimitIntoIndex.cpp
+  ReplaceEntriesWithObjectIteration.cpp
+  ReplaceEqualAttributeAccess.cpp
+  VectorIndex.cpp
+  PushFilterVectorIndex.cpp
+  Cluster.cpp
+  Graph.cpp
+  IResearchView.cpp
+  IndexNode.cpp
+  Join.cpp
+  LateMaterializedCommon.cpp
+  ReplaceFunctions.cpp
 )
