+++ arangod/Aql/Optimizer/Rule/Cluster.cpp
