+++ arangod/Aql/Optimizer/Rule/CollectWithIndex.cpp
