+++ arangod/Aql/Optimizer/Rule/Graph.cpp
@@ -28,7 +28,7 @@
 #include "Aql/ExecutionNode/TraversalNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Optimizer.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesGraph.h"
+#include "Aql/Optimizer/Rule/Graph.h"
 #include "Basics/StaticStrings.h"
 #include "Containers/SmallVector.h"
