+++ arangod/Aql/Optimizer/Rule/Graph.h
