+++ arangod/Aql/Optimizer/Rule/IndexNode.cpp
@@ -33,8 +33,8 @@
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/Expression.h"
 #include "Aql/Optimizer.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesIndexNode.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesLateMaterializedCommon.h"
+#include "Aql/Optimizer/Rule/IndexNode.h"
+#include "Aql/Optimizer/Rule/LateMaterializedCommon.h"
 #include "Aql/QueryContext.h"
 #include "Basics/AttributeNameParser.h"
 #include "Cluster/ServerState.h"
