+++ arangod/Aql/Optimizer/Rule/PropagateConstantAttributes.cpp
