+++ arangod/Aql/Optimizer/Rule/PushFilterVectorIndex.cpp
