+++ arangod/Aql/Optimizer/Rule/ReplaceEntriesWithObjectIteration.cpp
