+++ arangod/Aql/Optimizer/Rule/ReplaceEqualAttributeAccess.cpp
