+++ arangod/Aql/Optimizer/Rule/ReplaceFunctions.cpp
