+++ arangod/Aql/Optimizer/Rule/VectorIndex.cpp
