+++ arangod/Aql/OptimizerRulesFeature.cpp
@@ -24,9 +24,9 @@
 #include "OptimizerRulesFeature.h"
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Aql/Optimizer/Rule/EnumeratePathsFilter/EnumeratePathsFilter.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesGraph.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesIResearchView.h"
-#include "Aql/Optimizer/Rule/OptimizerRulesIndexNode.h"
+#include "Aql/Optimizer/Rule/Graph.h"
+#include "Aql/Optimizer/Rule/IResearchView.h"
+#include "Aql/Optimizer/Rule/IndexNode.h"
 #include "Aql/OptimizerRules.h"
 #include "Aql/OptimizerRulesOptionsProvider.h"
 #include "Basics/Exceptions.h"
